requirements.txt
docopt
Pillow
numpy
scikit-learn
